Merge pull request #4256 from RosettaCommons/roccomoretti/Wsuggest_override
Override Overdrive

C++11 added the override keyword for virtual functions which are a re-implementation of a base class's virtual function. This is helpful, because the compiler can check to make sure that the function is actually overriding the base class's function, rather than defining a new function with a different interface. In fact, modern C++ recommended practice is to use virtual only for the original (base class) version of a virtual function, and any reimplementation of the function should be marked with override.

This PR runs the clang-tidy modernize-use-override tool over the codebase to add override and convert virtual to override where warranted.

After this, `virtual` should really only be on functions which are the initial declaration of a virtual function, and `override` should be on all declarations of virtual functions which are in a subsequent child class.

Old repository SHA1: 2fbc75295dca398147a7ddbe7dce14e6fa9b00f3