Merge pull request #4276 from RosettaCommons/vmullig/fix_pose_from_sfr_builder
Fix a bug in the PoseFromSFRBuilder

The -in:Ctermini and -in:Ntermini options aren't working as expected. This ensures that, if the user uses these, they properly apply termini only to those chains that are listed.

This also adds a unit test for the desired functionality.

Old repository SHA1: 52692eb4e4d559958744a1d890acadc53b630a64