Merge pull request #4304 from RosettaCommons/jadolfbr/fix_insert_conf_by_jump
This is a simple change that clears the `GlycanTreeSet` and observers at the beginning of the `insert_conformation_by_jump` function, and then sets it up at the end.  This is due to the way this function deals with observers - instead of adding a residue, and then calling residue append, it adds ALL residues to the pose and THEN calls residue append for all the residues at the end.

Old repository SHA1: c0544d95c0f4110010c2fcbea66e45c4187b4600