Merge pull request #4311 from RosettaCommons/vmullig/make_orient_function_const
Identified a function that can and should be const.

This was a small mistake corrected in a much longer pull request which might not make its way into master for a while. I want to make sure that this change goes in now, since any changes to this function that would prevent it from being const would mess up what I'm doing in that branch.

Old repository SHA1: 61da2f786e1882934861d83f906ea33689ef732f