Merge pull request #4295 from RosettaCommons/vmullig/rocco_fix_to_hbond_geom
Picking Rocco's fix to hbonds_geom.cc out of roccomoretti/bugprone-integer-division_fix.

Description:

In pull request #4290, Rocco has identified a number of places in which we have likely been making division errors by assuming that an integer divided by an integer gives a floating-point number (which it does not). These are bugs that need to be fixed. Unfortunately, at least one likely has implications for scoring or minimization. This is the change most likely to cause unit, integration, and scientific test changes, so I think it makes sense to test it and merge it entirely separately. I want to be sure that this is a change that we can make without re-calibrating the whole scoring function.

Notes on the current scientific test failures and changes:

--    make_fragments, RosettaCM, and glycan_structure_prediction are all failures-to-run because they're not yet in the current master.
--    dock_glycans and mp_symdock fail in the same way as they do currently in master.
--    Looking at fast_relax_5iter (which passes), I notice no remarkable qualitative difference from current master. (There are no large changes in the scores or score ranges, for example).
--    Looking at antibody_snugdock (which passes), there are big jumps in the discrimination score, but I gather that this happens from run to run anyways. The plots look qualitatively similar.
--    The simple_cycpep_predict test does a lot of sampling. Again, the E vs. RMSD lots look qualitatively very similar.
--    All tests passing in master pass in this branch.

Based on the scientific test results (https://b3.graylab.jhu.edu/revision/commits/13892), I'm pretty convinced that this change to scoring is sufficiently benign that we don't need to worry about recalibrating everything or about it invalidating scientific performance.

Old repository SHA1: 91b4d2d1beee08f2fa6245a80e63a0a503e94307

Co-authored-by: Rocco Moretti <rmorettiase@gmail.com>