Merge pull request #4260 from RosettaCommons/JackMaguire/BetterAnnealer
Expanding user interface to the SequenceSymmetricAnnealer

We're having a discussion on Slack about how the SequenceSymmetricAnnealer I added in #4117 could be used for a wider set of design cases. My goal in this PR is to provide a residue selector interface to the annealer so that the user can define regions that need to be sequence-symmetric.

Old repository SHA1: b67bf8ccf699fe79f9a03f6fe221c45834dd2b8f