Merge pull request #4402 from RosettaCommons/roccomoretti/release_indication_in_version
Add backtrace to console output for developers

A desire was recently expressed to return the backtrace information to the console/tracer output for developers.

In order to provide this info for developers (for whom it may be useful) but not for users-at-large (for whom it probably isn't), this PR adds an additional version info feature for being a "public release". This will be true for the Academic and Commercial downloads of Rosetta, but will be false for any version of Rosetta directly compiled from Git. While the recently added -run:crash_to_console flag will control things if given, if the option is omitted, the behavior will be print the results on the console for git/developer releases, and to leave it out for the public releases.

In all cases, the ROSETTA_CRASH.log file is still generated.

Old repository SHA1: 7f2bd0108df8569443fde0c4e75002e83fdfc3a5