Merge pull request #4413 from RosettaCommons/roccomoretti/submodule_update_test
An initial attempt at a submodule up-to-date test.

The way it works is to look at the versions of the submodule for the current commit at the merge point. If there's a regression (that is, if master has submodule commits which the new commit doesn't), then the test will fail with a magenta-failure subtest for that module. (This is special cased for tests on master itself, which looks to make sure we haven't regressed on either of the parent commits of the master merge.)

I've also added a check to see if we're up-to-date with the most recent version of the submodule's master. That's not an error, though, it's just reported in the test log.

Old repository SHA1: e56969d9778b6c995e4ae9280b47adee51758c25