Merge pull request #4461 from RosettaCommons/vmullig/fix_maybe_uninitialized_errors
Fix a couple of compilation errors in gcc/7.

The gcc/7 compiler on Ubuntu 18 was complaining about uninitialized or potentially uninitalized variables in a devel protocol. This may be a false warning, since other compilers don't seem to complain, but initializing the vars shushes the warnings, and is unlikely to have any significant impact on performance given the complexity of the other stuff that these classes do.

Old repository SHA1: 531813ae494df1ec1ae67dd202ef04de5f53d4cc