Merge pull request #4460 from RosettaCommons/vmullig/multithreaded_scoring_incremental_merge2
Incremental merge 2 for multithreaded scoring

This PR relaxes some of the locking of the global options system.  I was overdoing it, and it was hurting multithreaded performance and performance scaling a bit.

(Note that the full locking can be restored by defining `OPTIONS_COLLECTION_FULL_THREAD_SAFETY`.  This is not defined by default.  When `MULTI_THREADED` is defined, as it is when building with `extras=cxx11thread`, only the necessary locking occurs.)

- [x] Pull request #4459 must be merged before this one.

@bcov77 

Old repository SHA1: f01b52ba4dead724a6a62fa5b312fd1b0ff3cdbb