Merge pull request #4476 from RosettaCommons/roccomoretti/fix_cppcheck_test
Attempt to fix the cppcheck test.

We're having issues with cryptic errors in the cppcheck test.

Part of that is that the cppcheck test script has the wrong option for grep in parsing out error lines, and part of it is that cppcheck updates (1.90 in particular) changed the behavior of cppcheck. Fix that and update the "known issues" list accordingly.

Old repository SHA1: 2d8a8e2046c7cb5453685af036327813055917c1