Merge pull request #4486 from RosettaCommons/roccomoretti/crash_report_server
Add script for reporting crashes to the crash report server.

For a while now Rosetta has been producing crash report logs. Part of the intent of having these would be to eventually collect those reports so we could see where Rosetta was having bugs or needed better error handling. -- The big limit to this so far being putting together the actual site to collect the data.

Luki offered to put together such a website, and has done an awesome job of it: https://crash.rosettacommons.org/ (Submission of crash reports is open to the public, but viewing the reports needs a Github login.) There's still some open questions about how we want to summarize/organize/report the crash reports, but the actual crash report collection should be feature complete, and we have basic crash report search and summary functionality.

This PR adds mention of the site into the crash report logs, as well as adding a script which can automatically collect and submit the crash reports.

I also addressed a non-crash which was being reported as the most prevalent crash. (The exception raised by JD2 to report an issue with a job.) Not only will this reduce noise on the crash report server, this has the added benefit of making the tracer output look nicer in the case of errors.

Old repository SHA1: 3387ec42a9999de15df75abf56c11a784705b6cf