Merge pull request #4501 from RosettaCommons/JackMaguire/FixThanhsBug
Handling edge cases for the SequenceSymmetricAnnealer

SequenceSymmetricAnnealer started out as a hack and slowly matured into a feature that was merged to master in #4117 and #4260. Some edge-case handling was missed during the de-hackifying process.

The branch name is misleading. This is not Thanh's bug - this is my bug that Thanh found.

Thanks Rocco for the review!

Old repository SHA1: 286deaae4986f74e96d1ffcbdda130ea3221b0be