Merge pull request #4509 from RosettaCommons/vmullig/fix_simple_cycpep_predict_bug
Fix a bug in simple_cycpep_predict with disulfides in MPI mode

The addition of the DISULFIDE variants to the native pose when predicting structures of disulfide-cyclized peptides was messed up, but only in MPI mode.  The issue was that the SimpleCycpepPredictApplication object is initialized differently in the non-MPI mode (reading directly from disk) or in MPI mode (one process reads from disk and broadcasts stuff to all other processes). The initialization of the native pose was depending on information that had not yet been received by all processes in the latter case (the sequence length).

TODO:
- [x] Check the `simple_cycpep_predict` scientific test (queued here: https://b3.graylab.jhu.edu/revision/commits/14718 ). --> Yes, this runs correctly still.

Old repository SHA1: 321d95b6c7a2f2df9ef12360afdaf74d170ef559