Merge pull request #4549 from RosettaCommons/vmullig/citation_manager_incremental_merge_1
Incremental merge #1 for the citation manager

This is a first incremental merge for the citation manager.  All of the functionality is there, now, and works for all JD2 apps and for RosettaScripts.  Future incremental merges will make this work for JD3 and PyRosetta, and will add additional citations.  This is part of the functionality implemented in pull request #3087.

Old repository SHA1: 623eb78fcf36e2cda51fda74306383c2fac8d6d7

Co-authored-by: Jack Maguire <jackmaguire1444@gmail.com>
Co-authored-by: Jared Adolf-Bryfogle <jadolfbr@gmail.com>
Co-authored-by: Labonte <JWLabonte@jhu.edu>
Co-authored-by: Shannon Smith <shannon.t.smith.1@vanderbilt.edu>
Co-authored-by: Brandon Frenz <brandon.frenz@gmail.com>
Co-authored-by: Vikram K. Mulligan <vmullig@uw.edu>
Co-authored-by: Vikram K. Mulligan <vmullig@u.washington.edu>