Merge pull request #4544 from RosettaCommons/vmullig/peptidestubmover_respect_ACE_and_NME
Feature addition for the PeptideStubMover

Currently, the `PeptideStubMover` can only append or prepend to naked termini or to the default terminal types.  However, it would be convenient to be able to append to a C-methylamidated residue or to prepend on an N-acetylated residue, preserving the mainchain dihedral values when we do so (since these terminal modifications are meant to emulate an amide bond to a preceding or following residue).  This pull request adds this ability.

This PR also makes a minor change to the `FoldTree` and `Edge` classes (adding  a new `Edge` constructor and a new `FoldTree::add_edge()` method for convenience, each of which takes slightly different parameters than the existing methods).  I've ensured that these are covered in the `FoldTree` unit tests.

Tasks:
- [x] Add the feature.
- [x] Unit-test it.
- [x] Debug it.
- [x] Beauty.
- [x] Documentation.

Old repository SHA1: 20a7d010d46d9d2e73c9a666feef3e437e94de03