Merge pull request #4550 from RosettaCommons/vmullig/no_cite_nodesign_packer_palette
Minor adjustment to the NoDesignPackerPalette

Tweak the citation registration for the NoDesignPackerPalette. Currently, this is being reported as an unpublished used module in any RosettaScripts script that uses FastRelax (which uses the NoDesignPackerPalette under the hood). This is not a sufficiently important use of the NoDesignPackerPalette to warrant pointing it out to the user. I'm moving the registration with the CitationManager to the parse_my_tag() function so that only explicitly-scripted NoDesignPackerPalettes trigger an unpublished module warning. (This should NOT be the general pattern followed for all unpublished modules).

Old repository SHA1: 5534d82a77eba5012185de83fa40d2a223152505