Merge pull request #4577 from RosettaCommons/roccomoretti/repack_wo_ligand_fix
Fix RepackWithoutLigand

Amanda Loshbaugh discovered that the repack without ligand functionality was actually redesigning in some cases. This is due to the DetectProteinLigandInterface actually skipping the interface detection and just going with a default all design run in the abscence of an option. This is suboptimal, so change things such that we always detect the interface and never design in the repack without ligand case.

Old repository SHA1: 696e766ff0b7fd72dca31013b6f5bc1b6e3abb16