Merge pull request #4475 from RosettaCommons/roccomoretti/rename_ReferenceCount
Rename ReferenceCount class to VirtualBase.

The ReferenceCount class has been a misnomer for several years now. It's purpose is no longer related to owning pointers or referencing counting, but now is simply to serve two purposes. The first is to serve as a common base class in Rosetta (such that we can store a generic ReferenceCountOP and then typesafely dynamic_cast it back to the true class), and the second is to ensure that all of the classes in a polymorphic hierarchy (i.e. one where a base class pointer/reference can point to a derived type) have a virtual destructor.

Old repository SHA1: 8b850656abf3f89abf728f2b3ee9b1fc5491bcb0