Merge pull request #4613 from RosettaCommons/vmullig/update_pymol_selection_metric
Making SelectedResiduesPyMOLMetric robust to PDB files that lack PDBInfo, and providing the option of outputting Rosetta residue numbering instead of PDB residue numbering.

The SelectedResiduesPyMOLMetric currently assumes that all poses have PDBInfo. This updates the metric to be compatible with poses that don't, and adds an option to allow a user to force Rosetta numbering as output instead of PDB numbering, if he or she so wishes.

Old repository SHA1: 18d66fe0101defde5a00609efbe61010b1898d34