Merge pull request #4533 from RosettaCommons/roccomoretti/parse_input_refpose
Address using the input pose as a reference pose in parse_my_tag.

This PR deals with those Movers and Filters which purposely use the input pose as a reference structure. The approach I'm taking is to add the input pose to the DataMap, but in a fashion where it's (sort-of) treated like another reference pose - in fact, those Movers/Filters which use the input pose but can't already take a reference pose have been updated such that they can take a regular (SavePoseMover) reference pose.

Information about how to update things for the new scheme are documented in the "Updating RosettaScripts" page on the documentation wiki.

Old repository SHA1: f6d4e41e24da062dec9d0341d1d393f7baa995a9