Merge pull request #4636 from RosettaCommons/roccomoretti/filter_value_metric
Add a SimpleMetric which wraps a Filter.

We're looking at converting the usage of Filters for calculating values to the SimpleMetric framework. In that conversion process, it would be really useful to have a RealMetric that's able to wrap a Filter, for backwards compatibility purposes. -- That is, if you have some framework which currently stores a filter and then uses the report_sm() functionality of that stored filter, you would want to convert it to store a RealMetric and use the value calculated by that real metric. But you also want to preserve backward compatibility in case people have old XMLs which are using and passing around Filters.

That's what the FilterValueMetric is for. It's just a RealMetric which can take an arbitrary Filter, call report_sm() on it, and use that value as if it was calculated by a RealMetric proper.

The intent is not to be used "earnestly", but mainly as a backward compatibility shim during the process of updating things to use the metrics framework.

Old repository SHA1: 43186e059c1952f0e65fda4d473c5e39370eace4