Merge pull request #4732 from RosettaCommons/roccomoretti/cached_type_fix
Fix CachableDataType name map.

There's been some recent additions to the Pose's CacheableDataType. The issue is that the enum->name map wasn't updated along with it. Curiously, though there's a check for this, it doesn't seem to be exercised by any of the integration or unit tests.

Add a unit test which will catch this situation in the future, and fix the current lack of enum->string mapping for the three entries which are lacking it.

Old repository SHA1: 235c3e71b2bf3776e18d22d940878fdb288d0fb3