Merge pull request #4769 from RosettaCommons/vmullig/peptide_pnear_vs_ic_scitest
Add a scientific test for the correlation between peptide computed PNear and experimentally-measured IC50.

The rigidity of a peptide macrocycle is a determinant of binding affinity. When peptides are optimized for favourable interactions with a target, computed PNear (or DeltaG_folding) values correlate well with experimentally-measured IC50 values. This correlation has improved as the Rosetta energy function has improved, from a vague correlation with talaris2013 and talaris2014 to a pretty good correlation that actually predicts rank order with ref2015. (This is likely due to the practice of training against physics-based fluid simulations used for ref2015.) It would be good to ensure that this trend continues for future versions of the energy function, so I'm adding a scientific test.

Old repository SHA1: 886e40432568875a572cbeec19d0b4242c5076c5