Merge pull request #4846 from RosettaCommons/roccomoretti/calculator_sm
Convert the CalculatorFilter to a CalculatorMetric

In light of the push to move from Filters to SimpleMetrics for value calculation, this PR adds a new RealMetric, the CalculatorMetric, which can take other RealMetrics and compute a value with them. This is very similar in intent and function to the CalculatorFilter. (In fact, the CalculatorFilter was re-implemented in terms of the CalculatorMetric, using the FilterValueMetric as a bridge.)

Also, because I could, you can now define the metrics being used for variables "in-line" under the CalculatorMetric tag. (The name parameter is not optional in this case.)

Old repository SHA1: dfbee7af523706bcc047f748854137066d90eebd