Possible fix to stochastic failure in HRFDyamicsEnergyTests (#4910)
The one thing that I can see is that there's no check that the data read from each line of the initialization file can actually be interpreted as an int and a float (which the code expects).  If this fails, the int and the float _should_ be set to 0 and 0.0, respectively, under the cxx11 standard, but earlier standards left the container variables unchanged.  I'm wondering if that's what we're seeing here, resulting in stochastic behaviour since the int and the float are uninitialized.

I'm adding a check that the line was properly parsed, and taking out the blank line in the config file.  I don't know that this will fix the stochastic failure, but it can't hurt.

Tasks:
- [x] Add the check.
- [x] Confirm that the unit test now fails consistently due to the blank line.
- [x] Remove the blank line in the config file.
- [x] Ensure that test passes. 
* Remove blank line in test.

* Update tools repo to point at latest master.

Old repository SHA1: a69d51d4b89e70701e22c8c95b6c13df85a3f23c