Merge pull request #4859 from RosettaCommons/vmullig/handle_disulfides_in_mutate_residue
Allow the MutateResidue mover to break disulfide bonds automatically.

It's a bit silly to have the MutateResidue mover leave a pose in an unstable, un-scorable state with a "dangling" disulfide partner (which still expects to be bonded to something) if you try to mutate a disulfide-bonded cysteine. This PR ensures that the mover automatically removes the disulfide partner's DISULFIDE variant type when it makes the mutation.  There is an option to restore the old behaviour in the rare cases in which a user might want that, but it's clearly marked as not recommended.

Old repository SHA1: 26e100a2fbe40d4bfb539294f836fbc130d66fb4