Merge pull request #4928 from RosettaCommons/vmullig/tidy_up_clang_tidy_warnings
Removing unused values flagged by clang-tidy.

Note that the operator() called here has no side-effects; its ONLY effect is to return a value that's never used, so it's safe to delete it.

Additional notes added by Rocco:

Looking at the code and #4890, it looks like the removal is fine. The computation that was previously done with these variables has been moved to the output_interface_deltas().

Presuming the component energies sum to the total scores, then the transformation should be valid. (As the difference of the sums of the components should be equal to the sums of the differences in components.)

... Well, with the caveat that you're either not using the normalization function, or that the normalization function is a linear (not affine) scaling, and the sum of normalized components is equal to the normalized sum of components. -- But looking through the current options for normalization factors, that's the case. (And the normalization option isn't used much, anyway.)

Old repository SHA1: 6dedbd4c240c738acaeac4698f33c74f6dd74076