Merge pull request #4938 from RosettaCommons/roccomoretti/grid_cache_size
Fix issue with memory usage in ligand docking.

There's been a number of reports of ligand docking eating up memory. Most of these can be fixed by setting -max_grid_cache_size

I don't think there's a reason not to have this set to a finite value, so I'm making that change to the default.

Additionally, I think the memory growth is in part due to a fix I added a while back which adds the grid center to the lookup hash string. Because the ligand center moves during docking, when the InterfaceCalculator goes to calculate final grid scores, it creates a new (used-only-once) grid set. What we can do is allow the GridManager to pull out a grid which is "close enough", which should pick up the original grid in many cases and thus not cause a grid regeneration and associate memory bloat/cache churn.

Old repository SHA1: 24d644b24f244cb64fbb1b59b0684e6340e3f9d7