Merge pull request #4856 from RosettaCommons/vmullig/move_cavity_volume_from_devel_to_protocols
Move CavityVolumeFilter from devel to protocols

The `CavityVolumeFilter` was used in Silva _et al._ (2019.) De novo design of potent and selective mimics of IL-2 and IL-15.  _Nature_ 565: 186–191, but the scripts in the supplement can not be run, and by extension the work that the paper describes is not reproducible, if this remains in devel.  At least one person has been asking about this on the Rosetta forums.

Tasks:
- [x] Move this filter to protocols from devel.
- [x] Move supporting class, the `CavityCalculator`, to protocols from devel.
- [x] Add citation info.  I'm going to count this as an unpublished module since Tom wasn't on the paper in which it first appeared, and it wasn't really described in that paper, and Tom might want to write a paper describing this.

Also:
- [x] Update the tools submodule to point at the latest master for the tools repo.

Old repository SHA1: 35b3a5cfadfc22354d034ad55367a6bc82b119b4