Merge pull request #4714 from RosettaCommons/roccomoretti/rdkit_submodule2
Add RDKit as a submodule to Rosetta.

While Rosetta has a bunch of functionality for macromolecule modeling, it's lacking functionality for small molecule modeling.

The RDKit (http://rdkit.org/) is a widely-used BSD-licensed library for chemoinformatics and analysis & manipulation of small molecules. While primarily used as a Python package, RDKit has a C++ library backing. This PR adds RDKit as a submodule in the external/ directory, and hooks the C++ library into Rosetta compilation. (Such that the files will be compiled with Rosetta.)

In addition to the library, included in this PR are two classes to facilitate transformation from a Rosetta ResidueType to an RDKit molecule and back again. It also includes a SimpleMetric and a Mover which can be used to run a number of chemoinformatic measures (such as logP, TPSA, etc.) over a specified ResidueType.

Other functionality such as rotamer library generation and ligand manipulation will be added in future PRs. (This PR is one in a series to merge my drug_design branch into master.)

Old repository SHA1: d13ff3f9e5963c36c92283ea727413bb79823312

Co-authored-by: Sergey Lyskov <sergey.lyskov@jhu.edu>